schoolconnect-mcp-server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "schoolconnect-mcp-server"
version = "1.0.0"
description = "MCP server providing AI assistants with school announcement and calendar tools"
authors = [
    {name = "SchoolConnect Team", email = "support@schoolconnect.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.0.0",
    "python-dotenv>=1.0.0",
    "airtable-python-wrapper>=0.15.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "python-dateutil>=2.8.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/vimarshsub/schoolconnect-mcp-server"
Repository = "https://github.com/vimarshsub/schoolconnect-mcp-server"
Issues = "https://github.com/vimarshsub/schoolconnect-mcp-server/issues"

[project.scripts]
schoolconnect-mcp-server = "server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

